Moves attack builder tests in specific package
